requirements.txt
elasticsearch==8.12.0
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.2.0
pandas==2.1.4
numpy==1.26.2
logging==0.4.9.6
asyncio==3.4.3
aiofiles==23.2.1
cryptography==41.0.8
httpx==0.26.0
pytest==7.4.3
pytest-asyncio==0.23.2
